More refactoring to reflect change in library naming scheme
Also some small amendments and comment additions/deletions ready for merging back into main branch
