+++ DmgLocationUtils/src/main/java/com/dmgdesignuk/locationutils/easylocationutility/LocationRequestCallback.java
@@ -3,7 +3,7 @@ package com.dmgdesignuk.locationutils.easylocationutility;
 import android.location.Location;
 
 
-public interface DeviceLocationCallback {
+public interface LocationRequestCallback {
 
     void onLocationResult(Location location);
     void onFailedRequest(String result);
